# swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Todolist App API",
    "description": "개인 인증 기반 할일 관리 웹 애플리케이션의 RESTful API 명세입니다.\n\n## 인증\n모든 할일 관련 API는 JWT Bearer 토큰 인증이 필요합니다. 로그인 후 발급된 Access Token을 `Authorization: Bearer <token>` 형식으로 헤더에 포함하여 요청하십시오.\n\n## 토큰 만료\nAccess Token의 유효기간은 **1시간**입니다. 만료된 토큰으로 요청 시 `401 Unauthorized`가 반환됩니다.\n\n## 응답 형식\n모든 응답은 JSON 형식으로 반환됩니다.\n\n- **성공 응답:** `{ \"success\": true, \"data\": { ... } }`\n- **목록 응답:** `{ \"success\": true, \"data\": [...], \"pagination\": { ... } }`\n- **에러 응답:** `{ \"success\": false, \"error\": { \"code\": \"...\", \"message\": \"...\" } }`\n\n## 할일 상태 산출 규칙\n할일 상태는 DB에 저장되지 않으며, `start_date`, `due_date`, `is_completed`, 현재 날짜로부터 동적으로 산출됩니다.\n\n| 상태 | 산출 조건 |\n|------|-----------|\n| `pending` (시작전) | `is_completed = false` AND 오늘 < `start_date` |\n| `in_progress` (진행중) | `is_completed = false` AND `start_date` <= 오늘 <= `due_date` |\n| `overdue` (완료 실패) | `is_completed = false` AND 오늘 > `due_date` |\n| `completed` (성공 완료) | `is_completed = true` |",
    "version": "1.0.0",
    "contact": {
      "name": "Todolist App"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "로컬 개발 서버"
    }
  ],
...
  "paths": {
...
    "/api/todos": {
      "post": {
        "tags": ["Todos"],
        "summary": "할일 등록",
        "description": "새로운 할일을 등록합니다. (UC-04)\n\n**유효성 검증 규칙:**\n- `title`: 1~200자, 공백만으로 구성 불가\n- `description`: 최대 2,000자 (선택)\n- `startDate`: ISO 8601 날짜 형식 (YYYY-MM-DD), 필수\n- `dueDate`: ISO 8601 날짜 형식 (YYYY-MM-DD), 필수, `startDate` 이상이어야 함\n\n`userId`는 JWT 토큰에서 자동 추출되며, `isCompleted` 초기값은 `false`입니다.",
        "operationId": "createTodo",
        "security": [
          { "BearerAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTodoRequest"
              },
              "example": {
                "title": "OpenAPI 명세 작성",
                "description": "Swagger 3.0 JSON 형식으로 전체 API 명세를 완성한다.",
                "startDate": "2026-04-01",
                "dueDate": "2026-04-03"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "할일 등록 성공. 생성된 할일 정보를 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Todo"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "userId": "550e8400-e29b-41d4-a716-446655440000",
                    "title": "OpenAPI 명세 작성",
                    "description": "Swagger 3.0 JSON 형식으로 전체 API 명세를 완성한다.",
                    "startDate": "2026-04-01",
                    "dueDate": "2026-04-03",
                    "isCompleted": false,
                    "status": "in_progress",
                    "createdAt": "2026-04-01T09:00:00.000Z",
                    "updatedAt": "2026-04-01T09:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "입력값 유효성 검증 실패.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "invalidTitle": {
                    "summary": "제목 유효성 오류",
                    "value": {
                      "success": false,
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "제목은 1자 이상 200자 이하이며, 공백만으로 구성될 수 없습니다."
                      }
                    }
                  },
                  "invalidDescription": {
                    "summary": "설명 길이 초과",
                    "value": {
                      "success": false,
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "설명은 최대 2,000자까지 입력할 수 있습니다."
                      }
                    }
                  },
                  "invalidDateFormat": {
                    "summary": "날짜 형식 오류",
                    "value": {
                      "success": false,
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "날짜는 ISO 8601 형식 (YYYY-MM-DD)으로 입력해 주세요."
                      }
                    }
                  },
                  "dueDateBeforeStartDate": {
                    "summary": "종료일이 시작일보다 이전",
                    "value": {
                      "success": false,
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "종료일은 시작일 이후여야 합니다."
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "get": {
        "tags": ["Todos"],
        "summary": "할일 목록 조회",
        "description": "현재 로그인한 사용자의 할일 목록을 조회합니다. (UC-05)\n\n본인 소유의 할일만 반환됩니다. 상태 필터, 정렬, 페이지네이션을 지원합니다.\n\n**페이지네이션:**\n- 기본 페이지 크기: 20건\n- 최대 페이지 크기: 100건 (초과 요청 시 100건으로 자동 제한)\n\n**`status` 필터 값:**\n- `pending`: 시작전 (is_completed=false AND 오늘 < startDate)\n- `in_progress`: 진행중 (is_completed=false AND startDate <= 오늘 <= dueDate)\n- `overdue`: 완료 실패 (is_completed=false AND 오늘 > dueDate)\n- `completed`: 성공 완료 (is_completed=true)\n- `closed`: 종료된 할일 (overdue + completed 모두 포함)\n- 미지정 시 전체 목록 반환",
        "operationId": "getTodos",
        "security": [
          { "BearerAuth": [] }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "할일 상태 필터. `closed` 선택 시 완료 실패(overdue)와 성공 완료(completed)를 모두 포함합니다.",
            "schema": {
              "type": "string",
              "enum": ["pending", "in_progress", "overdue", "completed", "closed"]
            },
            "example": "in_progress"
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "description": "정렬 기준 필드. 기본값은 `createdAt`입니다.",
            "schema": {
              "type": "string",
              "enum": ["startDate", "dueDate", "createdAt"],
              "default": "createdAt"
            },
            "example": "dueDate"
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "description": "정렬 방향. 기본값은 `desc`입니다.",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "desc"
            },
            "example": "asc"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "조회할 페이지 번호 (1부터 시작). 기본값은 `1`입니다.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "example": 1
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "description": "페이지당 건수. 기본값은 `20`이며, 최대 `100`건입니다. 100을 초과하면 100으로 자동 제한됩니다.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "example": 20
          }
        ],
        "responses": {
          "200": {
            "description": "할일 목록 조회 성공. 등록된 할일이 없으면 빈 배열과 totalCount 0을 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Todo"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/PaginationInfo"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "userId": "550e8400-e29b-41d4-a716-446655440000",
                      "title": "OpenAPI 명세 작성",
                      "description": "Swagger 3.0 JSON 형식으로 전체 API 명세를 완성한다.",
                      "startDate": "2026-04-01",
                      "dueDate": "2026-04-03",
                      "isCompleted": false,
                      "status": "in_progress",
                      "createdAt": "2026-04-01T09:00:00.000Z",
                      "updatedAt": "2026-04-01T09:00:00.000Z"
                    }
                  ],
                  "pagination": {
                    "page": 1,
                    "size": 20,
                    "totalCount": 1,
                    "totalPages": 1
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/todos/{id}": {
      "get": {
        "tags": ["Todos"],
        "summary": "할일 상세 조회",
        "description": "특정 할일의 상세 정보를 조회합니다. (UC-06)\n\n본인 소유의 할일만 조회할 수 있습니다. 타인 소유의 할일 조회 시 `403 Forbidden`을 반환합니다.",
        "operationId": "getTodoById",
        "security": [
          { "BearerAuth": [] }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TodoId"
          }
        ],
        "responses": {
          "200": {
            "description": "할일 상세 조회 성공.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Todo"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "userId": "550e8400-e29b-41d4-a716-446655440000",
                    "title": "OpenAPI 명세 작성",
                    "description": "Swagger 3.0 JSON 형식으로 전체 API 명세를 완성한다.",
                    "startDate": "2026-04-01",
                    "dueDate": "2026-04-03",
                    "isCompleted": false,
                    "status": "in_progress",
                    "createdAt": "2026-04-01T09:00:00.000Z",
                    "updatedAt": "2026-04-01T09:00:00.000Z"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "put": {
        "tags": ["Todos"],
        "summary": "할일 수정",
        "description": "특정 할일의 내용을 수정합니다. (UC-07)\n\n본인 소유의 할일만 수정할 수 있습니다. 요청 바디의 모든 필드를 포함하여 전송해야 합니다 (PUT 전체 교체 방식).\n\n**유효성 검증 규칙:**\n- `title`: 1~200자, 공백만으로 구성 불가\n- `description`: 최대 2,000자 (선택, null 허용)\n- `startDate`: ISO 8601 날짜 형식 (YYYY-MM-DD), 필수\n- `dueDate`: ISO 8601 날짜 형식 (YYYY-MM-DD), 필수, `startDate` 이상이어야 함\n\n수정 후 `updatedAt`이 현재 일시로 갱신됩니다.",
        "operationId": "updateTodo",
        "security": [
          { "BearerAuth": [] }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TodoId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTodoRequest"
              },
              "example": {
                "title": "OpenAPI 명세 작성 (수정)",
                "description": "Swagger 3.0 JSON 형식으로 전체 API 명세를 완성한다. 에러 응답 추가.",
                "startDate": "2026-04-01",
                "dueDate": "2026-04-04"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "할일 수정 성공. 수정된 할일 정보를 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Todo"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "userId": "550e8400-e29b-41d4-a716-446655440000",
                    "title": "OpenAPI 명세 작성 (수정)",
                    "description": "Swagger 3.0 JSON 형식으로 전체 API 명세를 완성한다. 에러 응답 추가.",
                    "startDate": "2026-04-01",
                    "dueDate": "2026-04-04",
                    "isCompleted": false,
                    "status": "in_progress",
                    "createdAt": "2026-04-01T09:00:00.000Z",
                    "updatedAt": "2026-04-01T10:30:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "입력값 유효성 검증 실패.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "invalidTitle": {
                    "summary": "제목 유효성 오류",
                    "value": {
                      "success": false,
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "제목은 1자 이상 200자 이하이며, 공백만으로 구성될 수 없습니다."
                      }
                    }
                  },
                  "dueDateBeforeStartDate": {
                    "summary": "종료일이 시작일보다 이전",
                    "value": {
                      "success": false,
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "종료일은 시작일 이후여야 합니다."
                      }
                    }
                  },
                  "invalidDateFormat": {
                    "summary": "날짜 형식 오류",
                    "value": {
                      "success": false,
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "날짜는 ISO 8601 형식 (YYYY-MM-DD)으로 입력해 주세요."
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "delete": {
        "tags": ["Todos"],
        "summary": "할일 삭제",
        "description": "특정 할일을 영구적으로 삭제합니다. (UC-09)\n\n본인 소유의 할일만 삭제할 수 있습니다. 삭제된 데이터는 복구할 수 없습니다.",
        "operationId": "deleteTodo",
        "security": [
          { "BearerAuth": [] }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TodoId"
          }
        ],
        "responses": {
          "204": {
            "description": "할일 삭제 성공. 응답 바디 없음."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/todos/{id}/complete": {
      "patch": {
        "tags": ["Todos"],
        "summary": "할일 완료 처리",
        "description": "특정 할일을 완료 상태로 변경합니다. (UC-08)\n\n`isCompleted`를 `true`로 업데이트하며, 할일 상태가 `completed`(성공 완료)로 변경됩니다. 본인 소유의 할일만 완료 처리할 수 있습니다.\n\n`updatedAt`이 현재 일시로 갱신됩니다.",
        "operationId": "completeTodo",
        "security": [
          { "BearerAuth": [] }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TodoId"
          }
        ],
        "responses": {
          "200": {
            "description": "할일 완료 처리 성공. 완료 처리된 할일 정보를 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Todo"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "userId": "550e8400-e29b-41d4-a716-446655440000",
                    "title": "OpenAPI 명세 작성",
                    "description": "Swagger 3.0 JSON 형식으로 전체 API 명세를 완성한다.",
                    "startDate": "2026-04-01",
                    "dueDate": "2026-04-03",
                    "isCompleted": true,
                    "status": "completed",
                    "createdAt": "2026-04-01T09:00:00.000Z",
                    "updatedAt": "2026-04-01T11:00:00.000Z"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/todos/{id}/incomplete": {
      "patch": {
        "tags": ["Todos"],
        "summary": "할일 완료 취소",
        "description": "완료 처리된 할일을 미완료 상태로 되돌립니다. (UC-10)\n\n`isCompleted`를 `false`로 업데이트하며, 할일 상태가 `start_date`, `due_date`, 현재 날짜 기준으로 재산출됩니다 (`pending`, `in_progress`, `overdue` 중 하나). 본인 소유의 할일만 완료 취소할 수 있습니다.\n\n`updatedAt`이 현재 일시로 갱신됩니다.",
        "operationId": "incompleteTodo",
        "security": [
          { "BearerAuth": [] }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TodoId"
          }
        ],
        "responses": {
          "200": {
            "description": "할일 완료 취소 성공. 완료 취소된 할일 정보를 반환합니다. 상태는 날짜 기준으로 재산출됩니다.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Todo"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "userId": "550e8400-e29b-41d4-a716-446655440000",
                    "title": "OpenAPI 명세 작성",
                    "description": "Swagger 3.0 JSON 형식으로 전체 API 명세를 완성한다.",
                    "startDate": "2026-04-01",
                    "dueDate": "2026-04-03",
                    "isCompleted": false,
                    "status": "in_progress",
                    "createdAt": "2026-04-01T09:00:00.000Z",
                    "updatedAt": "2026-04-01T12:00:00.000Z"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Access Token을 사용한 Bearer 인증입니다.\n\n로그인(`POST /api/auth/login`) 후 발급된 `token`을 `Authorization: Bearer <token>` 형식으로 헤더에 포함하십시오.\n\n**토큰 유효기간:** 1시간\n**토큰 저장 위치:** 클라이언트 메모리 (Zustand Store) - 새로고침 시 재로그인 필요"
      }
    },
    "parameters": {
      "TodoId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "할일 고유 식별자 (UUID v4)",
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
      }
    },
    "schemas": {
...
      "Todo": {
        "type": "object",
        "description": "할일 정보. `status` 필드는 DB에 저장되지 않으며 `isCompleted`, `startDate`, `dueDate`, 현재 날짜로부터 동적으로 산출됩니다.",
        "required": [
          "id",
          "userId",
          "title",
          "startDate",
          "dueDate",
          "isCompleted",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "할일 고유 식별자 (UUID v4)",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "할일 소유자의 사용자 ID (UUID v4)",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "할일 제목 (1~200자, 공백만으로 구성 불가)",
            "example": "OpenAPI 명세 작성"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true,
            "description": "할일 상세 설명 (선택, 최대 2,000자)",
            "example": "Swagger 3.0 JSON 형식으로 전체 API 명세를 완성한다."
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "description": "할일 시작일 (ISO 8601 날짜 형식, YYYY-MM-DD)",
            "example": "2026-04-01"
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "description": "할일 종료일 (ISO 8601 날짜 형식, YYYY-MM-DD). `startDate` 이상이어야 합니다.",
            "example": "2026-04-03"
          },
          "isCompleted": {
            "type": "boolean",
            "description": "완료 여부. 기본값은 `false`입니다.",
            "example": false
          },
          "status": {
            "type": "string",
            "enum": ["pending", "in_progress", "overdue", "completed"],
            "description": "할일 현재 상태 (파생 값, DB에 저장되지 않음).\n- `pending`: 시작전 (is_completed=false AND 오늘 < startDate)\n- `in_progress`: 진행중 (is_completed=false AND startDate <= 오늘 <= dueDate)\n- `overdue`: 완료 실패 (is_completed=false AND 오늘 > dueDate)\n- `completed`: 성공 완료 (is_completed=true)",
            "example": "in_progress"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "할일 생성 일시 (ISO 8601)",
            "example": "2026-04-01T09:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "할일 최종 수정 일시 (ISO 8601)",
            "example": "2026-04-01T09:00:00.000Z"
          }
        }
      },
      "PaginationInfo": {
        "type": "object",
        "description": "페이지네이션 정보 (Offset 기반).",
        "required": ["page", "size", "totalCount", "totalPages"],
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1,
            "description": "현재 페이지 번호 (1부터 시작)",
            "example": 1
          },
          "size": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100,
            "description": "현재 페이지 크기 (페이지당 건수)",
            "example": 20
          },
          "totalCount": {
            "type": "integer",
            "minimum": 0,
            "description": "전체 할일 건수 (필터 조건 적용 후)",
            "example": 42
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0,
            "description": "전체 페이지 수",
            "example": 3
          }
        }
      },
      "ApiError": {
        "type": "object",
        "description": "API 에러 상세 정보.",
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "DUPLICATE_EMAIL",
              "INTERNAL_ERROR"
            ],
            "description": "에러 코드.\n- `VALIDATION_ERROR` (400): 입력값 유효성 검증 실패\n- `UNAUTHORIZED` (401): 인증 토큰 없음 또는 만료\n- `FORBIDDEN` (403): 타인 데이터 접근 시도\n- `NOT_FOUND` (404): 리소스를 찾을 수 없음\n- `DUPLICATE_EMAIL` (409): 이메일 중복 (회원 가입 시)\n- `INTERNAL_ERROR` (500): 서버 내부 오류",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "description": "사용자에게 표시할 에러 메시지 (한국어)",
            "example": "종료일은 시작일 이후여야 합니다."
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "description": "성공 응답의 공통 래퍼.",
        "required": ["success"],
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청 처리 성공 여부",
            "enum": [true],
            "example": true
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "에러 응답의 표준 형식.",
        "required": ["success", "error"],
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청 처리 성공 여부",
            "enum": [false],
            "example": false
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        }
      },
...
      "CreateTodoRequest": {
        "type": "object",
        "description": "할일 등록 요청 바디.",
        "required": ["title", "startDate", "dueDate"],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "할일 제목 (1~200자, 공백만으로 구성 불가)",
            "example": "OpenAPI 명세 작성"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true,
            "description": "할일 상세 설명 (선택, 최대 2,000자)",
            "example": "Swagger 3.0 JSON 형식으로 전체 API 명세를 완성한다."
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "description": "할일 시작일 (ISO 8601 날짜 형식, YYYY-MM-DD)",
            "example": "2026-04-01"
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "description": "할일 종료일 (ISO 8601 날짜 형식, YYYY-MM-DD). `startDate` 이상이어야 합니다.",
            "example": "2026-04-03"
          }
        }
      },
      "UpdateTodoRequest": {
        "type": "object",
        "description": "할일 수정 요청 바디 (PUT 전체 교체 방식). 모든 필수 필드를 포함하여 전송해야 합니다.",
        "required": ["title", "startDate", "dueDate"],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "할일 제목 (1~200자, 공백만으로 구성 불가)",
            "example": "OpenAPI 명세 작성 (수정)"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true,
            "description": "할일 상세 설명 (선택, 최대 2,000자). null 전송 시 설명을 제거합니다.",
            "example": "Swagger 3.0 JSON 형식으로 전체 API 명세를 완성한다. 에러 응답 추가."
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "description": "할일 시작일 (ISO 8601 날짜 형식, YYYY-MM-DD)",
            "example": "2026-04-01"
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "description": "할일 종료일 (ISO 8601 날짜 형식, YYYY-MM-DD). `startDate` 이상이어야 합니다.",
            "example": "2026-04-04"
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "인증 토큰이 없거나 만료된 경우.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": {
                "code": "UNAUTHORIZED",
                "message": "인증이 필요합니다. 로그인 후 다시 시도해 주세요."
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "다른 사용자 소유의 리소스에 접근하려는 경우.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": {
                "code": "FORBIDDEN",
                "message": "해당 리소스에 대한 접근 권한이 없습니다."
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "요청한 리소스를 찾을 수 없는 경우.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": {
                "code": "NOT_FOUND",
                "message": "요청한 할일을 찾을 수 없습니다."
              }
            }
          }
        }
      },
      "InternalError": {
        "description": "서버 내부 오류가 발생한 경우.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": {
                "code": "INTERNAL_ERROR",
                "message": "서버 내부 오류가 발생했습니다. 잠시 후 다시 시도해 주세요."
              }
            }
          }
        }
      }
    }
  }
}